Extract 'cannot delete' modal into own component
