+++ src/modules/formBuilder/components/formTree/FormTreeLabel.tsx
@@ -1,4 +1,4 @@
-import { Alert, IconButton, List, ListItem, Typography } from '@mui/material';
+import { IconButton, Typography } from '@mui/material';
 import { Box, Stack, Theme } from '@mui/system';
 import { TreeItem2Label, UseTreeItem2Parameters } from '@mui/x-tree-view';
 import { useTreeItem2 } from '@mui/x-tree-view/useTreeItem2/useTreeItem2';
@@ -8,14 +8,13 @@ import { useFormContext, useFormState } from 'react-hook-form';
 import { FormTreeContext } from './FormTreeContext';
 import useUpdateFormStructure from './useUpdateFormStructure';
 import CommonMenuButton from '@/components/elements/CommonMenuButton';
-import ConfirmationDialog from '@/components/elements/ConfirmationDialog';
 import {
   ConditionalIcon,
   DownIcon,
   UpIcon,
 } from '@/components/elements/SemanticIcons';
 import { FORM_ITEM_PALETTE } from '@/modules/formBuilder/components/FormBuilderPalette';
-import { displayLabelForItem } from '@/modules/formBuilder/formBuilderUtil';
+import CannotDeleteItemModal from '@/modules/formBuilder/components/formTree/CannotDeleteItemModal';
 import {
   FormItemPaletteType,
   ItemDependents,
@@ -31,12 +30,6 @@ export interface FormTreeLabelProps
     Omit<UseTreeItem2Parameters, 'children'> {
   itemId: string;
 }
-
-const dependentLabelMap: Record<string, string> = {
-  autofillDependents: 'Items with autofill condition(s)',
-  enableWhenDependents: 'Items with visibility condition(s)',
-  boundDependents: 'Items with min/max bound(s)',
-};
 
 const FormTreeLabel: React.FC<FormTreeLabelProps> = ({
   id,
@@ -118,51 +111,13 @@ const FormTreeLabel: React.FC<FormTreeLabelProps> = ({
       }}
     >
       {!!itemDependents && (
-        <ConfirmationDialog
-          // todo @Martha - move into its own component?
-          open={!!itemDependents}
-          title='Cannot delete item'
-          onConfirm={() => setItemDependents(undefined)}
-          confirmText='Close'
-          loading={false}
-          hideCancelButton={true}
-        >
-          "{displayLabelForItem(item)}" cannot be deleted because it is
-          referenced elsewhere.
-          <Alert
-            color='error'
-            icon={false}
-            sx={{ my: 2, '& .MuiAlert-message': { width: '100%' } }}
-          >
-            {Object.entries(itemDependents).map(([key, val]) =>
-              val?.length > 0 ? (
-                <Box key={key}>
-                  {dependentLabelMap[key]}:
-                  <List sx={{ listStyleType: 'disc', py: 0 }}>
-                    {val.map((dep) => (
-                      <ListItem
-                        sx={{
-                          display: 'list-item',
-                          ml: 4,
-                          px: 0,
-                          maxWidth: 'calc(100% - 32px)',
-                        }}
-                        key={dep.linkId}
-                      >
-                        {ancestorLinkIdMap[dep.linkId].map(
-                          (ancestor) =>
-                            `"${displayLabelForItem(itemMap[ancestor])}" > `
-                        )}
-                        {`"${displayLabelForItem(dep)}"`}
-                      </ListItem>
-                    ))}
-                  </List>
-                </Box>
-              ) : null
-            )}
-          </Alert>
-          All references to this item must be removed before it can be deleted.
-        </ConfirmationDialog>
+        <CannotDeleteItemModal
+          item={item}
+          itemMap={itemMap}
+          itemDependents={itemDependents}
+          setItemDependents={setItemDependents}
+          ancestorLinkIdMap={ancestorLinkIdMap}
+        />
       )}
       {displayAttrs && (
         <Stack
